fix(Solid): updates Input styles
